// ChaggarCharts.Web/src/data/src/models/LoginModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ChaggarCharts
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface LoginModel
 */
export interface LoginModel {
    /**
     * 
     * @type {string}
     * @memberof LoginModel
     */
    username: string;
    /**
     * 
     * @type {string}
     * @memberof LoginModel
     */
    password: string;
}

export function LoginModelFromJSON(json: any): LoginModel {
    return LoginModelFromJSONTyped(json, false);
}

export function LoginModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginModel {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'username': json['username'],
        'password': json['password'],
    };
}

export function LoginModelToJSON(value?: LoginModel | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'username': value.username,
        'password': value.password,
    };
}